LayoutTests/imported/w3c/web-platform-tests/css/css-nesting/invalidation-002-expected.txt

Test passes if color is green.

FAIL CSS Selectors nested invalidation on changed child assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
